requirements.txt
aniso8601==7.0.0
click==8.1.3
Flask==2.1.2
Flask-GraphQL==2.0.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
promise==2.3
psycopg2==2.9.3
python-dotenv==0.20.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
zipp==3.8.0
